requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
jaydebeapi==1.2.3
JPype1==1.5.0
pydantic==2.6.4
pandas==2.2.2
openpyxl==3.1.2